refactor(ipc): IPC 拆薄为 Service 层 + 契约独立成 @meebox/ipc 包
ipc.ts 从 2449 行收薄至约 50 行，仅做上下文装配与各域 handler 注册；
业务实现下沉到 apps/desktop/src/main/services/：

- common/ 公共工具：broadcast / pr-lookup（收口 adapter 查找）/ mirror /
  comments-cache / usage；context.ts 聚合依赖与公共工具
- 跨域 service：run-queue（pr-agent run 队列）、agent-orchestrator（编排 + AutoPilot）
- 四域 handler：app / pr / config / agent，各自维护领域私有工具

IPC 契约从 packages/shared/src/ipc.ts 拆分迁入新内部包 @meebox/ipc，按业务领域
分文件（app/pr/config/agent/events/common）；renderer/preload/main 三侧改从
@meebox/ipc 导入。SyncProgressEvent 因被 repo-mirror 消费，归位到 @meebox/shared。

纯结构性重构，不改运行行为；lint/typecheck/test/build 全绿。
