+++ apps/desktop/src/renderer/src/components/ChatPane.tsx
@@ -4,11 +4,11 @@ import type { TFunction } from 'i18next';
 import ReactMarkdown from 'react-markdown';
 import remarkBreaks from 'remark-breaks';
 import remarkGfm from 'remark-gfm';
+import type { IpcChannels } from '@meebox/ipc';
 import type {
   AgentMessage,
   AgentStep,
   Finding,
-  IpcChannels,
   LocalPrStatus,
   PrAgentStatus,
   PrDocSectionKey,
