+++ apps/desktop/src/renderer/src/components/DiffSearchPanel.tsx
@@ -2,7 +2,7 @@ import { useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { TFunction } from 'i18next';
 import { editor as MonacoEditorNs } from 'monaco-editor';
-import type { DiffChangedFile } from '@meebox/shared';
+import type { DiffChangedFile } from '@meebox/ipc';
 import { invoke } from '../api';
 import { languageFor } from '../utils/language';
