+++ apps/desktop/typings/env.d.ts
@@ -5,7 +5,7 @@
  * vite/client 的全局类型通过 tsconfig.json 的 compilerOptions.types 引入；
  * 不再 /// reference，跟 logger 包共享 typings 的做法保持一致。
  */
-import type { IpcBridge } from '@meebox/shared';
+import type { IpcBridge } from '@meebox/ipc';
 
 declare global {
   interface Window {
