+++ packages/ipc/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@meebox/ipc",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Typed IPC 契约（renderer ↔ main）：按业务领域拆分的 IpcChannels / IpcEvents / IpcBridge",
+  "type": "module",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "lint": "eslint src --max-warnings=0"
+  },
+  "nx": {
+    "targets": {
+      "typecheck": { "cache": true },
+      "lint": { "cache": true }
+    }
+  },
+  "dependencies": {
+    "@meebox/shared": "*"
+  }
+}
