+++ packages/shared/src/index.ts
@@ -2,8 +2,8 @@ export * from './agent-contract.js';
 export * from './app-info.js';
 export * from './config.js';
 export * from './inline-comment-policy.js';
-export * from './ipc.js';
 export * from './language.js';
 export * from './platform.js';
 export * from './poller-contract.js';
 export * from './pr-agent-status.js';
+export * from './sync-progress.js';
